Move validation to separate file
